Update seeders functionality: by default no clearing for most seeders

+++ database/seeders/CategoriesTableSeeder.php
@@ -4,11 +4,10 @@ namespace Database\Seeders;
 
 use Domain\Products\Models\Category;
 use Domain\Seo\Models\Seo;
-use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 
-class CategoriesTableSeeder extends Seeder
+class CategoriesTableSeeder extends BaseSeeder
 {
     /**
      * Run the database seeds.
@@ -17,6 +16,10 @@ class CategoriesTableSeeder extends Seeder
      */
     public function run()
     {
+        if (!$this->shouldClearData()) {
+            return;
+        }
+
         DB::statement('SET FOREIGN_KEY_CHECKS=0;');
         Category::query()->truncate();
         DB::statement('SET FOREIGN_KEY_CHECKS=1;');
